refactor: Naming 변경 및 테스트 수정
